Add the ability to resolve event messages using an external database. (#2)
Events do not contain all the information required to analyse them. An event provider uses an external DLL to store the event message text, and the event only refers to it by event ID.

It is sometimes difficult to understand what the event is saying without access to this message. However if the evtx file was removed from its originating system (or the software generating the message was uninstalled) it is possible to lose access to the correct message.

This change allows us to export all the events from all providers on a system into an SQLite database. We can then use this database to resolve the correct message when parsing evtx files - thereby filling in the message data.